chore: Move month picker to usage
