+++ web_ui/src/pages/user-management/usage/month-picker/month-picker.module.scss
